// sbsp_frontend/src/types/Easing.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Easing = { "type": "linear" } | { "type": "inPowi", "intensity": number } | { "type": "outPowi", "intensity": number } | { "type": "inOutPowi", "intensity": number } | { "type": "inPowf", "intensity": number } | { "type": "outPowf", "intensity": number } | { "type": "inOutPowf", "intensity": number };